refactor(contracts): bound accounting types
Store assets, shares, and collateral as uint128, and store rates and timestamps as uint64.
These bounds guarantee that share conversions and interest calculations fit within uint256 arithmetic.

+++ contracts/test/SharesMath.t.sol
@@ -36,10 +36,10 @@ contract SmolphoPreviewHarness is Smolpho {
         uint256 totalBorrowAssets,
         uint256 totalBorrowShares
     ) external {
-        market.totalSupplyAssets = totalSupplyAssets;
-        market.totalSupplyShares = totalSupplyShares;
-        market.totalBorrowAssets = totalBorrowAssets;
-        market.totalBorrowShares = totalBorrowShares;
+        market.totalSupplyAssets = uint128(totalSupplyAssets);
+        market.totalSupplyShares = uint128(totalSupplyShares);
+        market.totalBorrowAssets = uint128(totalBorrowAssets);
+        market.totalBorrowShares = uint128(totalBorrowShares);
     }
 }
